+++ src/App.vue
@@ -1,22 +1,34 @@
 <!--
  * @Author: jack-pearson
  * @Date: 2021-11-22 17:31:39
- * @LastEditTime: 2022-01-10 18:00:14
+ * @LastEditTime: 2022-01-14 18:19:26
  * @LastEditors: jack-pearson
  * @FilePath: /yh-vue3-admin/src/App.vue
  * @Description: 
 -->
 
 <template>
-  <el-config-provider :locale="{ language: language }" :size="ElComponentSize">
+  <el-config-provider :locale="locale" :size="ElComponentSize">
     <router-view></router-view>
   </el-config-provider>
 </template>
 
 <script setup lang="ts">
 import { computed } from "vue";
+import ElementZhCn from "element-plus/lib/locale/lang/zh-cn";
+import ElementEn from "element-plus/lib/locale/lang/en";
 import { settingsStore } from "@/store";
 const settingsState = settingsStore();
-const ElComponentSize = computed(() => settingsState.ElComponentSize);
-const language = computed(() => settingsState.language);
+const ElComponentSize = computed(() => settingsState.config.ElComponentSize);
+const language = computed(() => settingsState.config.language);
+const locale = computed(() => {
+  switch (language.value) {
+    case "zh-cn":
+      return ElementZhCn;
+    case "en":
+      return ElementEn;
+    default:
+      return ElementZhCn;
+  }
+});
 </script>