+++ src/layout/navbar/index.vue
@@ -1,7 +1,7 @@
 <!--
  * @Author: jack-pearson
  * @Date: 2021-12-31 18:10:15
- * @LastEditTime: 2022-01-11 14:25:02
+ * @LastEditTime: 2022-01-13 18:20:46
  * @LastEditors: jack-pearson
  * @FilePath: /yh-vue3-admin/src/layout/navbar/index.vue
  * @Description: 
@@ -17,6 +17,7 @@
       <LayoutSearch />
       <LayoutZiTiSize />
       <LayoutLanguage />
+      <LayoutTheme />
       <LayoutUser />
     </div>
   </header>
@@ -31,6 +32,7 @@ import LayoutBreadcrumb from "./breadcrumb.vue";
 import LayoutSearch from "./search.vue";
 import LayoutZiTiSize from "./zitiSize.vue";
 import LayoutLanguage from "./language.vue";
+import LayoutTheme from "./theme.vue";
 import LayoutUser from "./user.vue";
 </script>
 <style lang="scss" scoped>