+++ src/layout/sidebar/SidebarItem.vue
@@ -1,13 +1,13 @@
 <!--
  * @Author: jack-pearson
  * @Date: 2021-12-30 15:18:38
- * @LastEditTime: 2022-01-04 15:41:46
+ * @LastEditTime: 2022-01-14 18:23:22
  * @LastEditors: jack-pearson
  * @FilePath: /yh-vue3-admin/src/layout/sidebar/SidebarItem.vue
  * @Description: 
 -->
 <template>
-  <div v-if="!isHide(route)" class="w-full sidebar-item" :class="{ 'is-collapsed': settingsConfig.isCollapsed }">
+  <div v-if="!isHide(route)" class="w-full sidebar-item" :class="{ 'is-collapsed': settingsConfig.config.isCollapsed }">
     <template v-if="!hasChildren(route)">
       <el-menu-item class="sub-menu select-none" :index="route.path">
         <template #title>
@@ -19,7 +19,7 @@
     <el-sub-menu v-else class="sub-menu select-none" :index="route.path" popper-append-to-body>
       <template #title>
         <svg-icon :name="route.meta.icon"></svg-icon>
-        <span v-show="!settingsConfig.isCollapsed" class="ml-3.5">{{ i18nRouter(route.meta.title) }}</span>
+        <span v-show="!settingsConfig.config.isCollapsed" class="ml-3.5">{{ i18nRouter(route.meta.title) }}</span>
       </template>
       <SidebarItem v-for="child in route.children" :key="child.id" :route="child" />
     </el-sub-menu>