+++ src/layout/sidebar/sidebarLogo.vue
@@ -1,7 +1,7 @@
 <!--
  * @Author: jack-pearson
  * @Date: 2021-12-31 13:51:06
- * @LastEditTime: 2022-01-11 10:37:24
+ * @LastEditTime: 2022-01-14 18:23:26
  * @LastEditors: jack-pearson
  * @FilePath: /yh-vue3-admin/src/layout/sidebar/sidebarLogo.vue
  * @Description: 
@@ -10,7 +10,7 @@
 <template>
   <div class="sidebar-logo w-full flex items-center justify-center h-auto p-1.5">
     <img :src="logo" alt="logo" class="logo w-12 h-12 cursor-pointer select-none" />
-    <div class="logo-title pl-2.5" :class="{ 'opacity-0': settingsConfig.isCollapsed }">{{ config.logoTitle }}</div>
+    <div class="logo-title pl-2.5" :class="{ 'opacity-0': settingsConfig.config.isCollapsed }">{{ config.logoTitle }}</div>
   </div>
 </template>
 <script setup lang="ts">