+++ src/types/store/router/index.d.ts
@@ -1,13 +1,11 @@
 /*
  * @Author: jack-pearson
  * @Date: 2021-12-14 14:07:07
- * @LastEditTime: 2022-01-11 16:44:28
+ * @LastEditTime: 2022-01-14 19:00:45
  * @LastEditors: jack-pearson
  * @FilePath: /yh-vue3-admin/src/types/store/router/index.d.ts
  * @Description:
  */
-export type DbBoolean = 0 | 1;
-
 export interface Menu {
   id?: number;
   parentId?: number;
@@ -16,11 +14,13 @@ export interface Menu {
   redirect: string;
   component: any;
   children: Menu[];
-  isHide: DbBoolean;
-  meta: {
-    title: string;
-    icon: string;
-    isKeepAlive: DbBoolean;
+  isHide?: boolean;
+  meta?: {
+    title?: string;
+    icon?: string;
+    isKeepAlive?: boolean;
+    isTagView?: boolean;
+    isAffix?: boolean;
   };
   sort?: number;
   createTime?: Date;