+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 <!--
  * @Author: jack-pearson
  * @Date: 2021-12-10 16:33:42
- * @LastEditTime: 2022-01-13 18:06:57
+ * @LastEditTime: 2022-01-14 19:03:39
  * @LastEditors: jack-pearson
  * @FilePath: /yh-vue3-admin/CHANGELOG.md
  * @Description: 
@@ -28,3 +28,7 @@
 `2022年01月13日`
 
 - 🐞 tagView 完善, 剩下刷新未做, 但是整体来说 tagView 做的很粗糙(后面等着重新设计)
+
+`2022年01月14日`
+
+- 🐞 tagView 完善, 侧边设置组件增加.修改 store 的代码, 扩充 menu 数据.